+++ base/exports.jl
@@ -657,6 +657,7 @@ export
     isequal,
     ismutable,
     isless,
+    isunordered,
     ifelse,
     objectid,
     sizeof,